requirements.txt
PyGithub==2.1.1
python-dotenv==1.0.0
pandas==2.1.1
pyyaml==6.0.1
click==8.1.7
